+++ test_gen/src/SendMessage.cpp
@@ -1,6 +1,6 @@
 #include "main.hpp"
 
-void SendMessage(redisContext *c, float val, std::string StreamName)
+void SendMessage(redisContext *c, double val, std::string StreamName)
 {
     redisReply *r = RedisCommand(c, "XADD %s * value %f", StreamName.c_str(), val);
     assertReplyType(c, r, REDIS_REPLY_STRING);
